refactor: move correlated message template
Moves the template from Java package "index" to "template".
